automatically get needed info from NDMP storage before job
automatically fetch the information about the devices in the
ndmp agents so that we have
- tape names and addresses
- slot mappings

This probably can be enhanced.

Unil now, the NDMP logs were always delivered to the user agent which
is not possible when a job runs, so now it is also possible to
log to job messages.