d_msg: is now able to filter for sourcefile with regular expression.
d_msg can filter the debug output based on a regex and only print the
messages that are generated by files that match to the regex.

This is implemented with the c++11 <regex> header and we have
gtest unit tests that verify the functionality.

We have to check if this can be compiled on all platforms.

What is still missing is to enable the setdebug command to set this
filter string.